+++ cfm-root/cfm-ejb/src/main/java/pl/mg/cfm/model/Car.java
@@ -21,7 +21,7 @@ public class Car implements Serializable {
 
 
     @Column(name = "distance")
-    private Long distance;
+    private Double distance;
 
     @Column(name = "latitude")
     private Double latitude;
@@ -45,11 +45,11 @@ public class Car implements Serializable {
         this.carPk = carPk;
     }*/
 
-    public Long getDistance() {
+    public Double getDistance() {
         return distance;
     }
 
-    public void setDistance(long distance) {
+    public void setDistance(double distance) {
         this.distance = distance;
     }
