+++ cfm-root/cfm-web-client/src/main/java/pl/mg/cfm/webclient/data/dao/CarMapper.java
@@ -13,7 +13,7 @@ public class CarMapper implements RowMapper<CarPojo> {
     public CarPojo mapRow(ResultSet rs, int rowNum) throws SQLException {
         CarPojo car = new CarPojo();
         car.setCarId(rs.getString(CFMCarDaoImplementation.CAR_ID_COL));
-        car.setDistance(rs.getLong(CFMCarDaoImplementation.CAR_DISTANCE_COL));
+        car.setDistance(rs.getDouble(CFMCarDaoImplementation.CAR_DISTANCE_COL));
         car.setLatitude(rs.getDouble(CFMCarDaoImplementation.CAR_LATITUDE_COL));
         car.setLongitude(rs.getDouble(CFMCarDaoImplementation.CAR_LONGITUDE_COL));
         car.setOwnerId(rs.getInt(CFMCarDaoImplementation.CAR_OWNER_ID_COL));
